+++ test/esm/run-as-esmodules.spec.js
@@ -1,0 +1,15 @@
+import { JapaneseDateConverter } from "../../dist/japanese-date-converter";
+
+describe("JapaneseDateConverter", () => {
+  test("can be executed from ES Modules", () => {
+    // in class or function
+    const inputValue = "R01/05/01";
+    const settings = {
+      format: "yyyy/MM/dd",
+    };
+    const converter = new JapaneseDateConverter({ inputValue, settings });
+    const convertedValue = converter.execute();
+
+    expect(convertedValue).toBe("2019/05/01");
+  });
+});