+++ webpack.common.config.js
@@ -40,5 +40,5 @@ module.exports = {
   resolve: {
     extensions: ["*", ".js", ".ts"],
   },
-  target: ["web", "es5"]
+  target: ["web", "es5"],
 };