+++ webpack.dev.config.js
@@ -1,7 +1,7 @@
-const { merge } = require('webpack-merge');
-const common = require('./webpack.common.config.js');
+const { merge } = require("webpack-merge");
+const common = require("./webpack.common.config.js");
 
 module.exports = merge(common, {
-  mode: 'development',
+  mode: "development",
   devtool: false,
 });